+++ Changes.md
@@ -7,6 +7,7 @@ Breaking Changes
 - CyclesOptions :
   - Removed `useFrameAsSeed` plug. The frame is now automatically used as the seed if `seed` is not set.
   - Removed all texture cache options. These had never been exposed in the UI because this never became an offical Cycles feature.
+  - Removed `cryptomatteAccurate`. This feature is no longer present in Cycles.
 
 1.4.0.0b1 (relative to 1.3.x.x)
 =========